requirements.txt
numpy==1.18
pandas==1.1
tqdm==4.41  
pillow==7.0
opencv-python==4.1.2.30
matplotlib==3.2.2
jupyter2